+++ crates/github-actions-models/src/dependabot/v2.rs
@@ -65,6 +65,11 @@ pub enum Registry {
         username: Option<String>,
         password: Option<String>,
     },
+    GoproxyServer {
+        url: String,
+        username: Option<String>,
+        password: Option<String>,
+    },
     HexOrganization {
         organization: String,
         key: Option<String>,